+++ src/components/Router/ServerRoutes.js
@@ -7,7 +7,7 @@ import ClonePage from "../../views/Pages/ClonePage"
 import CreateDBPage from "../../views/Pages/CreateDBPage"
 import { useAuth0 } from "../../react-auth0-spa";
 import PrivateRoute from './PrivateRoute';
-import {NoLoginLayout} from './PrivateRoute'
+import {NoPageLayout} from './PrivateRoute'
 
 
 export const ServerRoutes = () => {
@@ -28,7 +28,7 @@ export const CloneRoutes = () => {
     const { isAuthenticated, loginWithRedirect, loading } = useAuth0();
     let nogo = "Log in to Terminus Hub, where you can clone databases shared by others to your local machine"
 
-    if(!loading && !isAuthenticated) return <NoLoginLayout text={nogo} loginWithRedirect={loginWithRedirect}/>
+    if(!loading && !isAuthenticated) return <NoPageLayout text={nogo} loginWithRedirect={loginWithRedirect}/>
     return (
         <Switch>
             <Route key="specificclone" path={`${CLONE_DB_ROUTE}${SPECIFIC_DB_ROUTE}`}>